feat(units): allow decimal values for beds and baths count
- Change data type from integer to decimal for beds and baths count in model, migration and requests
- Update import service to handle decimal values without rounding
- Add new create and edit drawer components for units management
- Include cities data in unit controller for drawer components

+++ database/migrations/2025_08_15_155359_create_units_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('tenants')->nullable();
             $table->date('lease_start')->nullable();
             $table->date('lease_end')->nullable();
-            $table->integer('count_beds')->nullable();
-            $table->integer('count_baths')->nullable();
+            $table->decimal('count_beds', 3, 1)->nullable(); // 3 digits total, 1 decimal place (e.g., 10.5)
+            $table->decimal('count_baths', 3, 1)->nullable();
             $table->string('lease_status')->nullable();
             $table->decimal('monthly_rent', 15, 2)->nullable();
             $table->string('recurring_transaction')->nullable();
